# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an api gateway",
        "title": "BeatSync API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/",
    "paths": {
        "/config": {
            "get": {
                "description": "show service config data when the service environment set to debug mode",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "get config data on the debug mode",
                "operationId": "get_config",
                "responses": {
                    "200": {
                        "description": "Response data",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_Ramazon1227_BeatSync_api_http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/config.Config"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Ramazon1227_BeatSync_api_http.Response"
                        }
                    }
                }
            }
        },
...
        "/v1/analysis/{analysis_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve analysis data by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data"
                ],
                "summary": "Get Analysis By ID",
                "operationId": "get-analysis-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Analysis ID",
                        "name": "analysis_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.HRVAnalysisResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Ramazon1227_BeatSync_api_http.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_Ramazon1227_BeatSync_api_http.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Ramazon1227_BeatSync_api_http.Response"
                        }
                    }
                }
            }
        },
...
        "/v1/sensor-data": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Save sensor data for analysis",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data"
                ],
                "summary": "Save Sensor Data",
                "operationId": "save-sensor-data",
                "parameters": [
                    {
                        "description": "Sensor data",
                        "name": "sensor_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SensorData"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_Ramazon1227_BeatSync_api_http.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Ramazon1227_BeatSync_api_http.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Ramazon1227_BeatSync_api_http.Response"
                        }
                    }
                }
            }
        },
        "/v1/user-analysis": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve all analysis data for a specific user within a date range",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data"
                ],
                "summary": "Get User Analysis",
                "operationId": "get-user-analysis",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start date in YYYY-MM-DD format",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date in YYYY-MM-DD format",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.UserHRVAnalysisResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Ramazon1227_BeatSync_api_http.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Ramazon1227_BeatSync_api_http.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "config.Config": {
            "type": "object",
            "properties": {
                "defaultLimit": {
                    "type": "string"
                },
                "defaultOffset": {
                    "type": "string"
                },
                "environment": {
                    "description": "debug, test, release",
                    "type": "string"
                },
                "httpport": {
                    "type": "string"
                },
                "httpscheme": {
                    "type": "string"
                },
                "influxBucket": {
                    "type": "string"
                },
                "influxDatabase": {
                    "type": "string"
                },
                "influxOrg": {
                    "type": "string"
                },
                "influxToken": {
                    "type": "string"
                },
                "influxURL": {
                    "type": "string"
                },
                "passcodeLength": {
                    "type": "integer"
                },
                "passcodePool": {
                    "type": "string"
                },
                "secretKey": {
                    "type": "string"
                },
                "serviceHost": {
                    "type": "string"
                },
                "serviceName": {
                    "type": "string"
                },
                "smtpfrom": {
                    "type": "string"
                },
                "smtphost": {
                    "description": "Email Configuration",
                    "type": "string"
                },
                "smtppassword": {
                    "type": "string"
                },
                "smtpport": {
                    "type": "integer"
                },
                "smtpusername": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "github_com_Ramazon1227_BeatSync_api_http.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "description": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.HRVAnalysisResult": {
            "type": "object",
            "properties": {
                "analysis_id": {
                    "description": "Unique identifier for the analysis",
                    "type": "string"
                },
                "analysis_time": {
                    "description": "Timestamp of when the analysis was performed",
                    "type": "string"
                },
                "hf": {
                    "type": "number"
                },
                "lf": {
                    "type": "number"
                },
                "lf_hf_ratio": {
                    "type": "number"
                },
                "nn50": {
                    "type": "integer"
                },
                "pnn50": {
                    "type": "number"
                },
                "rmssd": {
                    "type": "number"
                },
                "sd1": {
                    "type": "number"
                },
                "sd2": {
                    "type": "number"
                },
                "sdnn": {
                    "type": "number"
                },
                "user_id": {
                    "type": "string"
                },
                "vlf": {
                    "type": "number"
                }
            }
        },
...
        "models.PPGData": {
            "type": "object",
            "required": [
                "timestamp",
                "value"
            ],
            "properties": {
                "timestamp": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "models.SensorData": {
            "type": "object",
            "required": [
                "data",
                "timestamp",
                "user_id"
            ],
            "properties": {
                "data": {
                    "description": "Heart rate value (e.g., beats per minute)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PPGData"
                    }
                },
                "device_id": {
                    "description": "Optional: Identifier for the specific device",
                    "type": "string"
                },
                "sensor_data_id": {
                    "description": "Unique identifier for the sensor data",
                    "type": "string"
                },
                "timestamp": {
                    "description": "Timestamp of the reading",
                    "type": "string"
                },
                "user_id": {
                    "description": "Unique identifier for the user",
                    "type": "string"
                }
            }
        },
...
        "models.UserHRVAnalysisResponse": {
            "type": "object",
            "properties": {
                "analysis": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.HRVAnalysisResult"
                    }
                },
                "count": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Enter the token with Bearer prefix, e.g. \"Bearer abcde12345\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}